The last checkin was broken in the case where class B inherited from class A, and class A had made itself immutable.
In that case, MMMC thought that the inlined constructor for A meant it
couldn't inline for B. This checkin fixes it to be much smarter. If
any parent is inlined, it checks to see if the inlined constructor's
class in the parent matches the constructor class that's trying to do
inlining. If they match, it goes ahead and inlines, otherwise it
doesn't.