Removed undocumented groups feature Moose::Exporter and updated tests accordingly.
The feature was good in spirit, but the implementation was very odd, and
basically undocumentable because of its complexity.

I moved some tests from the groups tests to
t/050_metaclasses/012_moose_exporter.t, since there were also useful tests for
prototype handling and with_meta mixed in with the groups tests.