When warning on mutable parents in make_immutable, ignore anon classes.
An anon parent class almost certainly comes from appyling base class roles
with Moose::Util::MetaRole, and the fact that said parent is still mutable is
not an error.

Of course, we could be wrong sometimes, but we're better off missing a warning
than spitting out a ton of useless ones.