Force parsing of blessed() as a function
In the generated constructor it was parsed and invoked as an indirect
method, which worked for classes that did not remove C<blessed> or
define it as an accessor.

This broke Search::GIN which had a class that used namespace::clean and
had a 'blessed' attribute which was incorrectly invoked on the class.