Refactored native trait accessors so they are done entirely in roles.
The class to do the inlining is generated based on the attribute's accessor &
delegation metaclasses, which means this works even when the attribute changes
those classes.

Added code to check if the attr's slot access can be inlined, and to use a
slower but working code path if it can't (see MX::AttributeTree for an example
of why this might be needed).