restore documentation for lazy_build and underscore (_) prefixed operators
It was removed from Moose::Manual::Attributes reasonably because it's
discouraged. It needs to be documented somewhere, as confound says in the
reference not the guide.

Also updated the link in Recipe 8 to point to the reference.
(cherry picked from commit 970a92fa56f1ea409c8d7c5428392479292fd8d4)