Merged CMOP into Moose
This involved tweaking the XS code a bit so that the Moose XS takes care of
booting the CMOP XS. It also meant changing Class::MOP to load Moose.so.

Most tests pass, except one which is probably related to the earlier
dzilification.

We will probably need an additional dzil plugin to customize the Makefile.PL
generation to do all the funky XS stuff that the hand-written one currently
does.