Make inline constraints safer ...
If we're defining a temp var - do this in a do {} block to limit its scope and
prevent it from conflicting with another constraint's var.

Don't use return in an inline constraint (this is super bad)

Don't use $_ to iterate over a collection of values in generated code, since
this could easily conflict with the surrounding code.