Fix is_class_loaded on blead
SvOK is only valid on scalars. It's pointless to call it on the result of
de-referencing an arbitrary scalar.

Blead recently made calling ->VERSION upgrade the scalar of the VERSION glob to
a version object. This broke us as SvOK(version_object) doesn't mean what we
thought it would.

Just dropping the entire SvROK branch and checking for any valid scalar in the
VERSION glob is the right thing to do here.