Test (that fails) for an edge case in role composition
When composing three roles, it is possible for the composite to end up with
both a conflict on a method and a method of that name. In most practical
applications, this doesn't really matter, since the conflict causes an
exception that must be resolved by the consumer.

However, it means that the composite reports bogus information when has_method
or get_method() is called.