Fix implementation of _STRINGLIKE0
This was checking for string overloading but only considered that valid if it
returned a non-zero length string, even though for plain old strings, an empty
string was accepted.

In practice, this didn't matter, because if this check failed, it would just
check to see if "$_[0] eq q{}", which stringified the object.