simplify and make $number_of_attributes
The concerning part is comparing these:

perl ./benchmarks/lotsa-classes.pl 1000 1
perl ./benchmarks/lotsa-classes.pl 1000 20

In the first case, all are in the "under a
few seconds" category.  In the second case
Moose moves into ~30s land while the others
pretty much stay in "under a few seconds".

This is kinda expected since Moose is inlining
constructors and accessors so that runtime can
be faster.  But in a sense, Moose isn't scaling
as well as the others.