Note: The following releases were mistakes and have been removed from PAUSE:
2.0803
2.1000
2.1001
2.1002

A force-reindexing has been performed to remove all traces of these releases.
The release commits remain, as a testament to how much one person can screw up
in just a few short minutes.

Git tags will remain in the repository in order to indicate that subsequent
2.10xx releases must start with 2.1003, as some users may already have up to
2.1003 installed on their systems.  However, the tags no longer point to their
original releases, but now point to this commit, for more self-documentation.