Applied patch from author of Module::Runtime I didn't test it But now Moose's tests are failed of Module::Runtime 0.014
From author:

Some of Moose's code that uses M:R is passing undef for the version
parameter of use_module() or use_package_optimistically(), when it should
pass no version parameter.  The new version of M:R handles this correctly,
treating it as a version number that turns out to be invalid, whereas the
old version mistook undef for a lack of parameter.  The bug is in Moose.