fix Class::MOP::Class::Immutable::Trait::_throw_exception (RT#96112)
Fixes breakage in MooseX::ABC where it was calling $meta->throw_error from
within an immutable class's 'new' method, and _throw_exception was being
called as an 'around' method modifier.

When a package  applied as a trait via Class::MOP::MiniTrait to a target
class, methods duplicating existing subs on the class will be added as an
'around' method modifier, while those that don't will be applied as plain
subs. This means that _throw_exception can get an unpredictable parameter list
depending on what it's applied to (not everything already has a
_throw_exception method).  To get around this, we simply define our own sub to
be used in this file only, and do not apply it to the target class with all
our other subs.