skip top frames in stack trace to fix memory leak
This fixes Moose::Exception leaking. The stack frames include all of the
arguments of each function call. Previously, some of the frames would be
method calls on the exception object, and so would include the object itself
in the arguments. This would create a reference loop, leaking the exception
object until process termination. Instead, we skip the stack frames related to
building the object, and only include those from ->new onward.

closes #81.