Remove logic for dead Test::Builder branch
These tests are for a branch of Test::Builder that has been abandoned.
Further they depend on the version of Test::Builder to do their thing. A
version of Test::Builder that is above the one in these tests, but does
not have the logic they use, will be relased (alpha) in the next week or
2.