Core overloading support for roles
This is more or less a straight port of MX::Role::WithOverloading.

I added some checks for overloading conflicts in MMR::Summation. However,
these need tests and I suspect that the implementation may be a bit off. Next
step is fixing that.

We now properly detect fallback & method conflicts in overloading setting
during role summation and when applying a role to another role.