list all runtime and configure prerequisites as authordeps
Part of Moose's documentation is built by introspecting the classes.
This means we need to be able to build and run Moose itself for a dzil
build to succeed.  This hasn't been a problem until now, because we
already have Moose installed as a prerequisite of Dist::Zilla.  But when
we add prereqs, the installed prereqs aren't enough.