Don't add command line option if compiler is cl
cl does not understand -Wdeclaration-after-statement. I am trying to
figure out if there is a corresponding warning option for cl, but, in
the mean time, not adding this command line option when the compiler
is cl will allow users to build Moose from a git checkout directory
using cl on Windows.