fix method cache generation when package contains blessed subs
When generating the method cache, the code was trying to call ->body on
it if it was an object.  The add_method code had previously been fixed
to only call ->body on meta objects rather than anything blessed.  The
XS code was only checking for objects being making that call.  This was
masked by method cache being kept up to date.

If the method cache is out of date though, this can crash.  The cache
can be intentionally reset, or we may not be able to track its status
properly (perl 5.8 without Class::C3::XS).

Fixes RT#113704