remove internal modules from no_index list
Moose has many modules that are only meant for use internally.  They
have been marked as no_index, preventing them from showing up in
02packages.  This serves nearly no purpose.  Although we don't want
people to depend on the modules, it would often work anyway because the
files still exist on disk.  It also means that the permissions for the
packages are not captured, allowing others to claim them, possibly
unintentionally.