remove test checking namespace::autoclean against Test::CleanNamespaces
The test was using namespace::autoclean to clean a namespace, then
checking that it was cleaned via Test::CleanNamespaces. This essentially
ends up only being a test of the external modules, not anything about
Moose itself. Both modules are using Moose to get the list of methods.
If that method list was wrong, the test would still pass because the
same list is being provided to both modules. And Moose's tests don't
need to include tests that are only of external modules.